change accrual sum on float

+++ internal/storage/order.go
@@ -8,7 +8,7 @@ import "gorm.io/gorm"
 type Order struct {
 	gorm.Model
 	OrderNumber string `gorm:"unique"`
-	Accrual     uint
+	Accrual     float32
 	UserID      uint
 	Status      string
 }
